files moved to expected folders
