+++ lib/assembly/object_file.rb
