+++ lib/assembly/object_file/version.rb
